Fixed get results and role permissions. Fixed some bugs.

+++ src/main/java/com/example/SoftbinatorProject/models/Ticket.java
@@ -21,7 +21,7 @@ public class Ticket {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    private Double amount;
+    private Integer amount;
 
     private Double price;
